refactor: :recycle: Change service module name from generic management to 'shops'
